chore: small change to custom_measure (doesn't work yet)

+++ vrtool/decision_making/measures/custom_measure.py
@@ -1,7 +1,5 @@
 import copy
 
-import numpy as np
-
 from vrtool.common.dike_traject_info import DikeTrajectInfo
 from vrtool.common.enums.mechanism_enum import MechanismEnum
 from vrtool.decision_making.measures.measure_protocol import MeasureProtocol
@@ -66,10 +64,11 @@ class CustomMeasure(MeasureProtocol):
                 ]
             )
 
+            # TODO VRTOOL-505: this code doesn't work yet
             mechanism_reliability = mechanism_reliability_collection.Reliability[
                 year_to_calculate
             ]
-            if np.int_(year_to_calculate) >= self.parameters["year"]:
+            if int(year_to_calculate) >= self.parameters["year"]:
                 if mechanism == MechanismEnum.OVERFLOW:
                     self._configure_overflow(mechanism_reliability)
                 elif mechanism == MechanismEnum.PIPING:
